♻️ Make `game` state module
