+++ src/states/index.ts
@@ -3,9 +3,8 @@ export {
   useIsMuted,
   showOppositePiecesValue,
   useShowOppositePieces,
-  drawValue,
-  setDraw,
 } from './settings';
+
 export {
   userNameValue,
   setUserName,
@@ -14,11 +13,7 @@ export {
   playerTurnValue,
   setPlayerTurn,
 } from './user';
-export {
-  roomIdValue,
-  setRoomId,
-  playerNamesValue,
-  setPlayerNames,
-  boardMapValue,
-  setBoardMap,
-} from './room';
+
+export { roomIdValue, setRoomId, playerNamesValue, setPlayerNames } from './room';
+
+export { drawValue, setDraw, boardMapValue, setBoardMap } from './game';
