+++ src/states/room.ts
@@ -15,12 +15,3 @@ export const setPlayerNames = (state: Pair<string>) => {
 };
 
 export const playerNamesValue = () => playerNames;
-
-/** A `Map` object that represents game board seen from the current user. */
-let boardMap: BoardMap = new Map();
-
-export const setBoardMap = (state: BoardMap) => {
-  boardMap = state;
-};
-
-export const boardMapValue = () => boardMap;
